pyproject.toml
[project]
name = "team-k-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "celery[redis]>=5.3.1",
    "debugpy>=1.8.14",
    "fastapi[standard]>=0.116.1",
    "google-cloud-storage>=3.2.0",
    "gunicorn>=23.0.0",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pika>=1.3.2",
    "pillow>=11.3.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "redis>=6.2.0",
    "ulid-py>=1.1.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.4",
]
